pyproject.toml
[project]
name = "tertuliando"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.0",
    "gradio>=6.5.1",
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
    "jupyter-dash>=0.4.2",
    "myfitnesspal>=2.1.2",
    "nbformat>=5.10.4",
    "numpy>=2.4.2",
    "openai>=2.21.0",
    "plotly>=6.5.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "ruff>=0.15.1",
    "scikit-learn>=1.8.0",
    "tiktoken>=0.12.0",
]
